Conflicts:
	Project290/Project290/Games/SuperPowerRobots/Entities/Weapon.cs

+++ Project290/Project290/Games/SuperPowerRobots/Entities/Bot.cs
@@ -47,10 +47,10 @@ namespace Project290.Games.SuperPowerRobots.Entities
 
             this.m_Weapons = new Weapon[4];
 
-            this.AddWeapon(0, "Gun", WeaponType.gun, (float) 100, (float) 10);
-            this.AddWeapon(1, "Gun", WeaponType.gun, (float)100, (float)10);
-            this.AddWeapon(2, "Shield", WeaponType.shield, (float) 500, (float) 50);
-            this.AddWeapon(3, "Gun", WeaponType.melee, (float)100, (float)0);
+            this.AddWeapon(0, "Gun", WeaponType.gun, 100f, 10f);
+            this.AddWeapon(1, "Gun", WeaponType.gun, 100f, 10f);
+            this.AddWeapon(2, "Shield", WeaponType.shield, 500f, 50f);
+            this.AddWeapon(3, "Gun", WeaponType.melee, 100f, 0f);
             
         }
